Minor fix due to error found locally regarding output file naming, etc

+++ src/crossref.py
@@ -91,28 +91,41 @@ def crossref(download, path):
     '''
     Download metadata from Crossref into a JSON file
     '''
-
-    paper = download.find('input', {'type': 'radio'})['value']
-    paper = paper.replace('https://hdl.handle.net/', 'https://doi.org/')
-    print(f'{current_timestamp()}: {paper}')
-    file_name = paper.replace('https://doi.org/', '')
-    file_name = file_name.replace('/', '-')
     try:
-        paper = crossref_commons.retrieval.get_publication_as_json(paper)
-        with open(f'{path}/{file_name}.json', 'w') as f:
-            json.dump(paper, f, indent=4)
-    except ValueError:
-        print(f'DOI {paper} does not exist')
+        paper = download.find('input', {'type': 'radio'})['value']
+        inputval = {
+            'https://hdl.handle.net/': 'https://doi.org/',
+            'http://hdl.handle.net/': 'https://doi.org/'
+        }
+        for key, value in inputval.items():
+            paper = paper.replace(key, 'https://doi.org/')
+        print(f'{current_timestamp()}: {paper}')
+        file_name = paper.replace('https://doi.org/', '')
+        file_name = file_name.replace('/', '-')
+        # no need to re-download existing file
+        if not os.path.exists(f'{path}/{file_name}.json'):
+            try:
+                paper = crossref_commons.retrieval.get_publication_as_json(paper)
+                sleep(3)
+                with open(f'{path}/{file_name}.json', 'w') as f:
+                    json.dump(paper, f, indent=4)
+            except ValueError:
+                print(f'DOI {paper} does not exist')
+    except AttributeError:
+        pass
 
 if __name__ == '__main__':
     journals = [
         # These are not an exhaustive list
-        's/oup/restud', # Oxford University Press
+        's/aea/aecrev', # American Economic Review
+        's/aea/jeclit', # Journal of Economic Literature
+        's/aea/jecper', # Journal of Economic Perspectives
+        's/eee/jfinec', # Journal of Financial Economics
         's/kap/jecgro', # Journal of Economic Growth
+        's/oup/qjecon', # The Quarterly Journal of Economics
+        's/oup/restud', # Oxford University Press
         's/oup/rfinst', # Review of Financial Studies
         's/ucp/jpolec', # Journal of Political Economy
-        's/eee/jfinec', # Journal of Financial Economics
-        's/oup/qjecon', # The Quarterly Journal of Economics
         's/wly/emetrp' # Econometrica
     ]
     PARSER = argparse.ArgumentParser()
